Core: Add `Service` Abstract Class
This commit adds the abstract `Service` class. This class is
foundational building block of `crossbill` and is a relative of, and
obviously inspired by Finagle's `Service`. At this time, this version
of `Service` only contains a `__call__` method which is capable of
asynchronously translating an instance of a `RequestType` into an
instance of a `ResponseType`.