Structure: Make Project Name `crossbill`
The name of this project is `crossbill`, even though the repository is
named `crossbillpy`. The `py` suffix is to disambiguate the project from
its Rust counterpart (which is where the name originated). It's expected
that at some point down the road the two will be intermingled.
Regardless, the Python version of `crossbill` should be named
`crossbill`.