Core: Add `Filter` Abstract Class
This commit adds the abstract `Filter` class. This class is one of the
foundational building blocks of `crossbill` and is a relative of, and
obviously inspired by Finagle's `Filter`. `Filter`s are intended to be
composed together and then composed on top of `Service`s to form a
`Request`/`Response` chain.

The `Filter` class diverges from its inspiration in that this `Filter`
acts more like a `SimpleFilter`, where the `Request` and `Response`
types do not change.

Additionally, none of the plumbing is in place yet to easily compose
`Filter`s together, nor compose `Filter`s and `Service`s.