Structure: Mark Project as Private
This commit modifies the `pyprojec.toml` file to mark the project as
private. The "Private :: Do Not Upload" classifier will prevent the
project from being uploaded to PyPI. This is temporary as the initial
API is put in place and the project's settings are rapidly changing.
That should change in the near future.