Structure: Tell Pyright where to Find Dependencies
This commit modifies the `tool.pyright` settings within the
`pyproject.toml` file to tell `pyright` about the `uv` virtual
environment (`.venv`). Prior to this change, in the test code, `pyright`
is unable to determine where the `import pytest` import is coming from.
Afterwards, this is fine, and no longer seen by `pyright` as an error.