Core: Add `Pipeline` Class
A `Pipeline` is intended for use in building a server, where `bytes` are
processed, converted to a `Request`, that `Request` is processed by a
`Service` which returns a `Response`, and then that `Response` is
converted back into `bytes`. A `Pipeline` is responsible for performing
those transformations given the correct pieces.