Test: Add `SpamService` for Testing
Using the `EchoService` for testing is helpful, but doing so means that
it's not clear whether data is making it all the way to the
`EchoService` or something else is intercepting it first and returning
it back as is. This commit adds a `SpamService` that ignores the
`StringRequest` data sent to it, and only responds with a
`StringResponse` with a value of "spam". Swapping out the `EchoService`
for the `SpamService` can provide assurance that a `Request` is hitting
  a `Service` as intended.