Core: Add `Closable` Protocol
This library is starting to take some shape, and with that, we'll need
to be responsible in our resource handling. Ensuring that resources are
properly `close`d is a big part of that. To that end, this commit
introduces a `Closable` protocol that classes should start implementing.

As an example, at some point in the future, `Service` should implement
`Closable`.