Transport: Add `Address` Representation
This commit encapsulates a `host` and a `port` as an `Address`.
`Address`es will likely become more complex over time and the
representation of other versions beyond `host` and `port` will be
possible. So, let's not start passing `host`s and `port`s all throughout
the code when the expectation is that support will exist for more than
that later.