Core: Add `PipelineFactory` Abstract Class
This commit adds a `PipelineFactory`, which is a factory class that is
capable of taking in a `Service` and returns a `Pipeline`. This is
useful for when development should be done in terms of a `Service`, that
`Service` is passed to another class, and behind the scenes that
`Service` is used to create a `Pipeline`.