Server: Add `Server` Class
A `Server` is capable of handling `Request`s and returning `Response`s
based on an `Address`. After creation, when calling `serve` on a
`Server`, it will listen on the provided `Address`. At that point,
`Client`s can connect to it, sending it `Request`s and having
`Response`s returned, based on the `Service` provided to the `serve`
method. It will remain in this state until `close` is called on the
`Server`.