Structure: Add Ability for Ruff to Sort Imports
Imports for this project have been a bit of a mess so far. It's time to
rectify that via tooling, in particular via `ruff`. This commit adds the
"I" value to `extend-select` for `ruff`.

extend-select: A list of rule codes or prefixes to enable, in addition to
those specified by select.

"I" is for `isort`.