bytes: Add `BytesRequest` Class
This commit creates the initial `crossbill.bytes` package and adds a
`BytesRequest` within it. `BytesRequest` is more intended to be a
structural element that is internal to `crossbill`. However, it's
entirely possible that dealing with `Request`s and `Response`s via
`BytesRequest` and `BytesResponse` would be very helpful for initial
development or protocol evolution.