Server: Modify `serve` Functionality for `Server` to Run Forever
This commit changes the behavior of `Server`'s `serve` method to handle
the creation of the underlying `asyncio` `Server` *and* to have that
`Server` `serve_forever`. Doing so requires a bounce through the event
loop without waiting on the finishing of that task. Instead `serve` now
returns based on `start_serving` being satisfied.