Structure: Add `iperf3` as a Dev Dependency
This commit adds `iperf3` as a development dependency. This initial
iteration of this library was built quickly, focused on the APIs and
usability. As such, there are significant gaps in the reliability of the
library that need to be hardened. Being able to handle a benchmarking
tool is a start.