Scripts: Add `benchmark_server` Script
This commit adds a script that can run a `StringEchoService` as a
`Server` or a version where every response is delayed by a fractional
amount. The initial set of numbers for the standard `StringEchoService`
are above expectations, so I'm not going to start publishing them until
I believe in their accuracy.