Transport: Close `StreamWriter` When Bridge is Done
This commit modifies the `__call__` method of a `Bridge` such that it
now calls `close` on the `StreamWriter` and also awaits the
`wait_closed` method of that `StreamWriter`.

Not doing so previously was a mistake, and although the code worked fine
on Python 3.13, it did not work correctly on 3.12 or 3.11. Making this
change paves the way for `crossbill` to work on those versions of
Python.