Structure: Set Pyright Python Version to 3.11
This commit sets the version that `pyright` should use for syntax
analysis to 3.11. That is the minimum version that we plan on supporting
at this time, and so code should work with that version of python. We
are already using some features, for example `asyncio.timeout` that were
added in 3.11, so supporting a lower version would require code changes.