Http: Add `HttpResponseCodec`
This commit adds an initial version of an `HttpResponseCodec`. It's
fairly crude and will grow more robust over time. The main idea for now
is to get it in place in order to complete the initial versions of an
`HttpClient` and `HttpServer`, and from there we're off to the races.